// packages/dnb-eufemia/src/components/upload/UploadFileInput.d.ts
import React from 'react';
import type { UploadFile } from './types';
declare const UploadFileInput: ({ children, disabled, files, ...rest }: {
    children?: React.ReactNode;
    files?: UploadFile[];
    disabled?: boolean;
}) => import("react/jsx-runtime").JSX.Element;
export default UploadFileInput;